# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SERMACROPS EDI Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List EDI documents
      parameters:
        - name: partnerId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: transactionSet
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of EDI documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EdiDocument"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create/upload an EDI document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EdiDocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocument"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get an EDI document with full detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: EDI document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocumentDetail"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update an EDI document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EdiDocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocument"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete an EDI document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /documents/{id}/validate:
    post:
      operationId: validateDocument
      tags: [documents]
      summary: Validate an EDI document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
  /documents/{id}/translate:
    post:
      operationId: translateDocument
      tags: [documents]
      summary: Translate an EDI document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Translation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationLog"
  /documents/{id}/send:
    post:
      operationId: sendDocument
      tags: [documents]
      summary: Send an EDI document to the trading partner via AS2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transmission log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransmissionLog"
  /documents/{id}/resend:
    post:
      operationId: resendDocument
      tags: [documents]
      summary: Resend a failed document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transmission log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransmissionLog"
  /documents/{id}/history:
    get:
      operationId: getDocumentHistory
      tags: [documents]
      summary: Get status history for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentStatusHistory"
  /documents/{id}/progress:
    get:
      operationId: getDocumentProgress
      tags: [documents]
      summary: Get EDI transaction progress flow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Progress flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentProgress"
  /documents/receive:
    post:
      operationId: receiveDocument
      tags: [documents]
      summary: Simulate receiving an inbound EDI document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveDocumentInput"
      responses:
        "201":
          description: Received document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdiDocument"

  # ─── Validation Results ───────────────────────────────────────────────────────
  /validation-results:
    get:
      operationId: listValidationResults
      tags: [documents]
      summary: List validation results
      parameters:
        - name: documentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Validation results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResult"

  # ─── Translation Logs ─────────────────────────────────────────────────────────
  /translation-logs:
    get:
      operationId: listTranslationLogs
      tags: [documents]
      summary: List translation logs
      parameters:
        - name: documentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Translation logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranslationLog"

  # ─── Transmission Logs ────────────────────────────────────────────────────────
  /transmission-logs:
    get:
      operationId: listTransmissionLogs
      tags: [documents]
      summary: List transmission logs
      parameters:
        - name: documentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transmission logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransmissionLog"

  # ─── Notifications ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    EdiDocument:
      type: object
      properties:
        id:
          type: string
        partnerId:
          type: string
        partnerName:
          type: ["string", "null"]
        endpointId:
          type: ["string", "null"]
        transactionSet:
          type: string
          enum: ["850", "855", "856", "810", "204", "990"]
        controlNumber:
          type: string
        status:
          type: string
          enum: [Draft, Validated, Sent, Received, Acknowledged, Failed]
        direction:
          type: string
          enum: [inbound, outbound]
        rawContent:
          type: ["string", "null"]
        parsedContent:
          type: ["object", "null"]
        validationStatus:
          type: string
          enum: [pending, passed, failed]
        isAcknowledged:
          type: boolean
        sentAt:
          type: ["string", "null"]
        receivedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - partnerId
        - transactionSet
        - controlNumber
        - status
        - direction
        - validationStatus
        - isAcknowledged
        - createdAt
        - updatedAt
    EdiDocumentDetail:
      type: object
      properties:
        id:
          type: string
        partnerId:
          type: string
        partnerName:
          type: ["string", "null"]
        endpointId:
          type: ["string", "null"]
        transactionSet:
          type: string
        controlNumber:
          type: string
        status:
          type: string
        direction:
          type: string
        rawContent:
          type: ["string", "null"]
        parsedContent:
          type: ["object", "null"]
        validationStatus:
          type: string
        isAcknowledged:
          type: boolean
        sentAt:
          type: ["string", "null"]
        receivedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/DocumentStatusHistory"
        validationResults:
          type: array
          items:
            $ref: "#/components/schemas/ValidationResult"
        translationLogs:
          type: array
          items:
            $ref: "#/components/schemas/TranslationLog"
        transmissionLogs:
          type: array
          items:
            $ref: "#/components/schemas/TransmissionLog"
      required:
        - id
        - partnerId
        - transactionSet
        - controlNumber
        - status
        - direction
        - validationStatus
        - isAcknowledged
        - createdAt
        - updatedAt
        - statusHistory
        - validationResults
        - translationLogs
        - transmissionLogs
    EdiDocumentInput:
      type: object
      properties:
        partnerId:
          type: string
        endpointId:
          type: string
        transactionSet:
          type: string
          enum: ["850", "855", "856", "810", "204", "990"]
        direction:
          type: string
          enum: [inbound, outbound]
        rawContent:
          type: string
      required:
        - partnerId
        - transactionSet
        - direction
    EdiDocumentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [Draft, Validated, Sent, Received, Acknowledged, Failed]
        rawContent:
          type: string
    ReceiveDocumentInput:
      type: object
      properties:
        partnerId:
          type: string
        endpointId:
          type: string
        transactionSet:
          type: string
          enum: ["850", "855", "856", "810", "204", "990"]
        rawContent:
          type: string
      required:
        - partnerId
        - transactionSet
        - rawContent

    # ─── Status History ─────────────────────────────────────────────────────────
    DocumentStatusHistory:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        status:
          type: string
          enum: [Draft, Validated, Sent, Received, Acknowledged, Failed]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - documentId
        - status
        - createdAt

    # ─── Validation Results ─────────────────────────────────────────────────────
    ValidationResult:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        isValid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
        warnings:
          type: array
          items:
            type: string
        validatedAt:
          type: string
      required:
        - id
        - documentId
        - isValid
        - errors
        - warnings
        - validatedAt
    ValidationError:
      type: object
      properties:
        segment:
          type: string
        field:
          type: ["string", "null"]
        code:
          type: string
        message:
          type: string
      required:
        - segment
        - code
        - message

    # ─── Translation Logs ───────────────────────────────────────────────────────
    TranslationLog:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        direction:
          type: string
          enum: [edi_to_json, json_to_edi]
        status:
          type: string
          enum: [success, failed]
        inputFormat:
          type: string
        outputFormat:
          type: string
        parsedSegments:
          type: integer
        errorMessage:
          type: ["string", "null"]
        translatedAt:
          type: string
      required:
        - id
        - documentId
        - direction
        - status
        - inputFormat
        - outputFormat
        - parsedSegments
        - translatedAt

    # ─── Transmission Logs ──────────────────────────────────────────────────────
    TransmissionLog:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        partnerId:
          type: string
        endpointUrl:
          type: string
        protocol:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        httpStatusCode:
          type: ["integer", "null"]
        responseMessage:
          type: ["string", "null"]
        as2MessageId:
          type: ["string", "null"]
        transmittedAt:
          type: string
      required:
        - id
        - documentId
        - partnerId
        - endpointUrl
        - protocol
        - status
        - transmittedAt

    # ─── Notifications ──────────────────────────────────────────────────────────
...
    DocumentProgress:
      type: object
      properties:
        documentId:
          type: string
        transactionSet:
          type: string
        flowType:
          type: string
          enum: [procurement, logistics]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ProgressStep"
        currentStep:
          type: integer
        totalSteps:
          type: integer
      required:
        - documentId
        - transactionSet
        - flowType
        - steps
        - currentStep
        - totalSteps
    ProgressStep:
      type: object
      properties:
        stepNumber:
          type: integer
        transactionSet:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [completed, current, pending, failed]
        documentId:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
      required:
        - stepNumber
        - transactionSet
        - label
        - status